Fixed zero velocity swipe
First met this issue on the Xperia M device (Android 4.1.2). Sometimes 'swipe' event didn't occur. This happens because Xperia M is dispatched 'touchmove' event so fast that on the slightest moving of finger 3 'touchmove's occurs with the same coordinates.
So, I think we should store the 'lastMove' coordinates only when the last 'touchmove' coordinates has at least 1px delta from the 'preLastMove' coordinates.